+++ src/main/java/com/aviad/coupons/entities/CategoryEntity.java
@@ -11,7 +11,7 @@ public class CategoryEntity {
 
     @Id
     @GeneratedValue
-    private short id;
+    private int id;
 
     @Column(name = "name", nullable = false, unique = true)
     private String name;
@@ -27,11 +27,11 @@ public class CategoryEntity {
         this.name = category.getName();
     }
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(short id) {
+    public void setId(int id) {
         this.id = id;
     }
